# openapi/blobyard-core.openapi.yaml
openapi: 3.1.0
info:
  title: Blob Yard Core API
  version: 1.0.0
  description:
    The stable API for durable object storage and scoped access. Stored files do not expire when a
    transfer grant or public capability expires.
servers:
  - url: https://api.blobyard.com/v1
security:
  - bearerAuth: []
paths:
...
  /yards:
    get:
      operationId: listWebYards
      description: Lists active Web Yards for one workspace project.
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/ListWebYardsSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - yard
          - list
      x-blobyard-mcp:
        tool: blobyard_list_web_yards
      parameters:
        - in: query
          name: project
          required: true
          schema:
            $ref: ./blobyard-shared.openapi.yaml#/components/schemas/Slug
        - in: query
          name: workspace
          required: true
          schema:
            $ref: ./blobyard-shared.openapi.yaml#/components/schemas/Slug
...
  /yards/sessions:
    get:
      operationId: listYardSessions
      description: Lists browser sessions for one Yard without exposing token material.
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/ListYardSessionsSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - yard-sessions
          - list
      x-blobyard-mcp:
        tool: blobyard_list_yard_sessions
      parameters:
        - in: query
          name: yardId
          required: true
          schema:
            $ref: ./blobyard-shared.openapi.yaml#/components/schemas/Identifier
...
  /users:
    get:
      operationId: listLocalUsers
      x-blobyard-deployments:
        - self-hosted
      description:
        Lists local users in one workspace with active sign-in key prefixes. Human sessions only.
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/ListLocalUsersSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - users
          - list
      x-blobyard-mcp:
        tool: blobyard_list_local_users
      parameters:
        - in: query
          name: workspace
          required: true
          schema:
            $ref: ./blobyard-shared.openapi.yaml#/components/schemas/Slug
    post:
      operationId: createLocalUser
      x-blobyard-deployments:
        - self-hosted
      description: Creates one local user and returns the raw sign-in key once. Human sessions only.
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/CreateLocalUserSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - users
          - create
      x-blobyard-mcp:
        excluded: Raw sign-in keys stay outside model context.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./blobyard-shared.openapi.yaml#/components/schemas/CreateLocalUserRequest
  /users/reset-key:
    post:
      operationId: resetLocalUserLoginKey
      x-blobyard-deployments:
        - self-hosted
      description:
        Revokes every active sign-in key for one user and returns the replacement key once. Human
        sessions only.
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/ResetLocalUserLoginKeySuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - users
          - reset-key
      x-blobyard-mcp:
        excluded: Raw sign-in keys stay outside model context.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./blobyard-shared.openapi.yaml#/components/schemas/ResetLocalUserLoginKeyRequest
  /users/deactivate:
    post:
      operationId: deactivateLocalUser
      x-blobyard-deployments:
        - self-hosted
      description:
        Deactivates one local user and revokes every active sign-in key. Human sessions only.
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/DeactivateLocalUserSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - users
          - deactivate
      x-blobyard-mcp:
        tool: blobyard_deactivate_local_user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./blobyard-shared.openapi.yaml#/components/schemas/DeactivateLocalUserRequest
...
  /cli/sessions:
    get:
      operationId: listCliSessions
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/ListCliSessionsSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - sessions
          - list
      x-blobyard-mcp:
        tool: blobyard_list_cli_sessions
...
components:
  securitySchemes:
    bearerAuth:
      $ref: ./blobyard-shared.openapi.yaml#/components/securitySchemes/bearerAuth
    githubOidc:
      $ref: ./blobyard-shared.openapi.yaml#/components/securitySchemes/githubOidc
...